+++ build_modules/lib/src/modules.dart
@@ -96,6 +96,13 @@ class Module {
   @JsonKey(name: 'm', nullable: true, defaultValue: false)
   final bool isMissing;
 
+  /// Whether or not this module is supported for [platform].
+  ///
+  /// Compilers can use this to either silently skip compilation of this module
+  /// or throw early errors or warnings.
+  ///
+  /// Modules are allowed to exist even if they aren't supported, which can help
+  /// with discovering root causes of incompatibility.
   @JsonKey(name: 'is', nullable: false)
   final bool isSupported;